refactor(PaymentClient): 결제 클라이언트 추상화
