+++ src/main/java/roomescape/domain/payment/PaymentClient.java
@@ -1,0 +1,8 @@
+package roomescape.domain.payment;
+
+import roomescape.service.dto.PaymentRequest;
+
+public interface PaymentClient {
+
+    void requestApproval(PaymentRequest request);
+}
