+++ src/main/java/roomescape/service/ReservationService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRepository;
+import roomescape.domain.payment.PaymentClient;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.ReservationRepository;
 import roomescape.domain.reservation.Waiting;
@@ -19,7 +20,6 @@ import roomescape.domain.reservation.dto.ReservationReadOnly;
 import roomescape.domain.reservation.slot.ReservationSlot;
 import roomescape.exception.AuthorizationException;
 import roomescape.exception.RoomEscapeBusinessException;
-import roomescape.infrastructure.PaymentClient;
 import roomescape.service.dto.LoginMember;
 import roomescape.service.dto.PaymentRequest;
 import roomescape.service.dto.ReservationBookedResponse;
